requirements.txt
fastapi
uvicorn
httpx
redis
pydantic
python-dotenv
typer
crewai
psycopg2-binary
requests
pyyaml

